need tests

+++ ex01/iter.hpp
@@ -4,9 +4,9 @@
 # include <iostream>
 
 template<typename T>
-void iter(T *  tab, size_t tabsize, void(*func)(T&))
+void iter(T*  tab, size_t tabsize, void(*func)(T&))
 {
-    for (int i = 0; i < tabsize; i++)
+    for (size_t i = 0; i < tabsize; i++)
     {
         func(tab[i]);
     }
